Factor out backend description code from `node.script`.
