+++ flocker/node/script.py
@@ -14,7 +14,7 @@ import yaml
 
 from jsonschema import FormatChecker, Draft4Validator
 
-from pyrsistent import PClass, field, PMap, pmap, pvector, pset_field
+from pyrsistent import PClass, field, PMap, pmap
 
 from eliot import ActionType, fields
 
@@ -25,15 +25,12 @@ from twisted.python.usage import Options, UsageError
 from twisted.internet.ssl import Certificate
 from twisted.internet import reactor  # pylint: disable=unused-import
 from twisted.internet.defer import succeed
-from twisted.python.constants import Names, NamedConstant
 
-from ..volume.filesystems import zfs
-from ..volume.service import (
-    VolumeService, DEFAULT_CONFIG_PATH, FLOCKER_MOUNTPOINT, FLOCKER_POOL)
 
 from ..common.script import (
     ICommandLineScript,
     flocker_standard_options, FlockerScriptRunner, main_for_service)
+from flocker.common.plugin import PluginNotFound
 from . import P2PManifestationDeployer, ApplicationNodeDeployer
 from ._loop import AgentLoopService
 from .exceptions import StorageInitializationError
@@ -44,13 +41,10 @@ from .diagnostics import (
 from .agents.blockdevice import (
     BlockDeviceDeployer, ProcessLifetimeCache,
 )
-from .agents.loopback import (
-    LoopbackBlockDeviceAPI,
-)
-from .agents.cinder import cinder_from_configuration
-from .agents.ebs import aws_from_configuration
 from ..ca import ControlServicePolicy, NodeCredential
 from ..common._era import get_era
+
+from .backends import DeployerType, get_backend
 
 __all__ = [
     "flocker_dataset_agent_main",
@@ -364,123 +358,6 @@ def get_configuration(options):
     return configuration
 
 
-def _zfs_storagepool(
-        reactor, pool=FLOCKER_POOL, mount_root=None, volume_config_path=None):
-    """
-    Create a ``VolumeService`` with a ``zfs.StoragePool``.
-
-    :param pool: The name of the ZFS storage pool to use.
-    :param bytes mount_root: The path to the directory where ZFS filesystems
-        will be mounted.
-    :param bytes volume_config_path: The path to the volume service's
-        configuration file.
-
-    :return: The ``VolumeService``, started.
-    """
-    if mount_root is None:
-        mount_root = FLOCKER_MOUNTPOINT
-    else:
-        mount_root = FilePath(mount_root)
-    if volume_config_path is None:
-        config_path = DEFAULT_CONFIG_PATH
-    else:
-        config_path = FilePath(volume_config_path)
-
-    pool = zfs.StoragePool(
-        reactor=reactor, name=pool, mount_root=mount_root,
-    )
-    api = VolumeService(
-        config_path=config_path,
-        pool=pool,
-        reactor=reactor,
-    )
-    api.startService()
-    return api
-
-
-class DeployerType(Names):
-    """
-    References to the different ``IDeployer`` implementations that are
-    available.
-
-    :ivar p2p: The "peer-to-peer" deployer - suitable for use with system like
-        ZFS where nodes interact directly with each other for data movement.
-    :ivar block: The Infrastructure-as-a-Service deployer - suitable for use
-        with system like EBS where volumes can be attached to nodes as block
-        devices and then detached (and then re-attached to other nodes).
-    """
-    p2p = NamedConstant()
-    block = NamedConstant()
-
-
-class BackendDescription(PClass):
-    """
-    Represent one kind of storage backend we might be able to use.
-
-    :ivar name: The human-meaningful name of this storage backend.
-    :ivar needs_reactor: A flag which indicates whether this backend's API
-        factory needs to have a reactor passed to it.
-    :ivar needs_cluster_id: A flag which indicates whether this backend's API
-        factory needs to have the cluster's unique identifier passed to it.
-    :ivar api_factory: An object which can be called with some simple
-        configuration data and which returns the API object implementing this
-        storage backend.
-    :ivar required_config: A set of the dataset configuration keys
-        required to initialize this backend.
-    :type required_config: ``PSet`` of ``unicode``
-    :ivar deployer_type: A constant from ``DeployerType`` indicating which kind
-        of ``IDeployer`` the API object returned by ``api_factory`` is usable
-        with.
-    """
-    name = field(type=unicode, mandatory=True)
-    needs_reactor = field(type=bool, mandatory=True)
-    # XXX Eventually everyone will take cluster_id so we will throw this flag
-    # out.
-    needs_cluster_id = field(type=bool, mandatory=True)
-    # Config "dataset" keys required to initialize this backend.
-    required_config = pset_field(unicode)
-    api_factory = field(mandatory=True)
-    deployer_type = field(
-        mandatory=True,
-        invariant=lambda value: (
-            value in DeployerType.iterconstants(), "Unknown deployer_type"
-        ),
-    )
-
-# These structures should be created dynamically to handle plug-ins
-_DEFAULT_BACKENDS = [
-    # P2PManifestationDeployer doesn't currently know anything about
-    # cluster_uuid.  It probably should so that it can make sure it
-    # only talks to other nodes in the same cluster (maybe the
-    # authentication layer would mostly handle this but maybe not if
-    # you're slightly careless with credentials - also ZFS backend
-    # doesn't use TLS yet).
-    BackendDescription(
-        name=u"zfs", needs_reactor=True, needs_cluster_id=False,
-        api_factory=_zfs_storagepool, deployer_type=DeployerType.p2p,
-    ),
-    BackendDescription(
-        name=u"loopback", needs_reactor=False, needs_cluster_id=False,
-        # XXX compute_instance_id is the wrong type
-        api_factory=LoopbackBlockDeviceAPI.from_path,
-        deployer_type=DeployerType.block,
-    ),
-    BackendDescription(
-        name=u"openstack", needs_reactor=False, needs_cluster_id=True,
-        api_factory=cinder_from_configuration,
-        deployer_type=DeployerType.block,
-        required_config={u"region"},
-    ),
-    BackendDescription(
-        name=u"aws", needs_reactor=False, needs_cluster_id=True,
-        api_factory=aws_from_configuration,
-        deployer_type=DeployerType.block,
-        required_config={
-            u"region", u"zone", u"access_key_id", u"secret_access_key",
-        },
-    ),
-]
-
 _DEFAULT_DEPLOYERS = {
     DeployerType.p2p: lambda api, **kw:
         P2PManifestationDeployer(volume_service=api, **kw),
@@ -489,28 +366,6 @@ _DEFAULT_DEPLOYERS = {
                             _underlying_blockdevice_api=api,
                             **kw),
 }
-
-
-def get_backend(backend_name, backends=_DEFAULT_BACKENDS):
-    """
-    Find the backend in ``backends`` that matches the one named by
-    ``backend_name``. If not found then attempt is made to load it as
-    plugin.
-
-    :param backend_name: The name of the backend.
-    :param backends: Collection of `BackendDescription`` instances.
-
-    :raise ValueError: If ``backend_name`` doesn't match any known backend.
-    :return: The matching ``BackendDescription``.
-    """
-    from flocker.common.plugin import get_plugin, PluginNotFound
-    try:
-        return get_plugin(backend_name, backends, "FLOCKER_BACKEND")
-    except PluginNotFound:
-        raise ValueError(
-            "'{!s}' is neither a built-in backend nor a 3rd party "
-            "module.".format(backend_name),
-        )
 
 
 def get_api(backend, api_args, reactor, cluster_id):
@@ -549,8 +404,6 @@ def get_api(backend, api_args, reactor, cluster_id):
 
 class AgentService(PClass):
     """
-    :ivar backends: ``BackendDescription`` instances describing how to use each
-        available storage backend.
     :ivar deployers: Factories to create ``IDeployer`` providers given an API
         object and some extra keyword arguments.  Keyed on a value from
         ``DeployerType``.
@@ -563,9 +416,6 @@ class AgentService(PClass):
     :ivar get_external_ip: Typically ``_get_external_ip``, but
         overrideable for tests.
     """
-    backends = field(
-        factory=pvector, initial=_DEFAULT_BACKENDS, mandatory=True,
-    )
     deployers = field(factory=pmap, initial=_DEFAULT_DEPLOYERS, mandatory=True)
     reactor = field(initial=reactor, mandatory=True)
 
@@ -627,7 +477,13 @@ class AgentService(PClass):
         :raise ValueError: If ``backend_name`` doesn't match any known backend.
         :return: The matching ``BackendDescription``.
         """
-        return get_backend(self.backend_name, self.backends)
+        try:
+            return get_backend(self.backend_name)
+        except PluginNotFound:
+            raise ValueError(
+                "'{!s}' is neither a built-in backend nor a 3rd party "
+                "module.".format(self.backend_name),
+            )
 
     # Needs tests: FLOC-1964.
     def get_tls_context(self):
